[feat] 공통 컴포넌트 수정
