+++ app/src/main/java/org/sopt/certi/core/component/section/CertInfoSection.kt
@@ -1,0 +1,41 @@
+package org.sopt.certi.core.component.section
+
+import androidx.annotation.DrawableRes
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import org.sopt.certi.core.util.screenWidthDp
+import org.sopt.certi.ui.theme.CertiTheme
+
+@Composable
+fun CertInfoSection(
+    @DrawableRes iconRes: Int,
+    testInfo: String,
+    modifier: Modifier = Modifier,
+    iconColor: Color = CertiTheme.colors.gray300
+) {
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(screenWidthDp(4.dp)),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Icon(
+            imageVector = ImageVector.vectorResource(iconRes),
+            contentDescription = null,
+            tint = iconColor
+        )
+        Text(
+            text = testInfo,
+            style = CertiTheme.typography.caption.regular_14,
+            color = CertiTheme.colors.black
+        )
+    }
+}
